client/requirements.txt
Flask == 1.1.2
Flask_SQLAlchemy == 2.5.1
Pillow == 8.2.0
numpy == 1.19.5
requests == 2.25.1
tensorflow == 2.5.0rc0
